baseai/subai/triaport/logistics_intelligence.py
import numpy as np

# Define the main class for the Triaport Logistics Intelligence module
class TriaportLogisticsIntelligence:
    def __init__(self):
        self.data_processor = DataProcessor()

    def analyze_logistics_data(self, data):
        processed_data = self.data_processor.process(data)
        analysis_result = self.perform_analysis(processed_data)
        return analysis_result

    def perform_analysis(self, data):
        # Example analysis logic
        mean_data = np.mean(data)
        return {'mean': mean_data}

# Helper class for data processing
class DataProcessor:
    def process(self, data):
        return np.array(data) * 1.1  # Example processing step
